template/requirements.txt
asgiref==3.2.10
autopep8==1.5.2
boto3==1.12.46
botocore==1.15.46
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-admin-sortable2==0.7.7
django-allauth==0.42.0
django-debug-toolbar==2.2
django-extensions==3.0.8
django-filter==2.3.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-storages==1.9.1
django-tables2==2.3.1
django-taggit==1.3.0
django-tinymce==3.0.2
djangorestframework==3.11.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.9.5
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
